backend/requirements.txt
fastapi
uvicorn
pillow
pandas
python-multipart
openpyxl
